requirements.txt
ftfy
gradio
openvino==2024.1
pillow
pymilvus
regex
tqdm
uuid
